Rewrites write_modpac() completely to parcel out reading/writing.
This introduces a new function, index_repository(), that constructs
the contents of the index from existing files and updates, and moves
all actual state changes to write_modpac().

In addition, both archiving features and hardlinking features are now
optional, and there is a flag for keeping only the latest version
(CRAN-style).